9180 Move EmploymentHodFormatters objects into companion objects. Still need to remove trait
