+++ app/uk/gov/hmrc/tai/model/domain/AnnualAccount.scala
@@ -16,8 +16,11 @@
 
 package uk.gov.hmrc.tai.model.domain
 
-import play.api.libs.json.{Format, Json}
-import uk.gov.hmrc.tai.model.tai.TaxYear
+import play.api.libs.json.{Format, JsArray, JsSuccess, JsValue, Json, Reads}
+import uk.gov.hmrc.tai.model.domain.EndOfTaxYearUpdate.endOfTaxYearUpdateHodReads
+import uk.gov.hmrc.tai.model.domain.Payment.paymentHodReads
+import uk.gov.hmrc.tai.model.tai.{JsonExtra, TaxYear}
+import uk.gov.hmrc.tai.model.tai.TaxYear.taxYearHodReads
 
 case class AnnualAccount(
   sequenceNumber: Int,
@@ -37,5 +40,47 @@ object AnnualAccount {
 
   def apply(sequenceNumber: Int, taxYear: TaxYear, rtiStatus: RealTimeStatus): AnnualAccount =
     AnnualAccount(sequenceNumber, taxYear, rtiStatus, Nil, Nil)
+
+  val annualAccountHodReads: Reads[Seq[AnnualAccount]] = (json: JsValue) => {
+
+    val employments: Seq[JsValue] = (json \ "individual" \ "employments" \ "employment").validate[JsArray] match {
+      case JsSuccess(arr, _) => arr.value.toSeq
+      case _                 => Nil
+    }
+
+    JsSuccess(employments.map { emp =>
+      val sequenceNumber = (emp \ "sequenceNumber").as[Int]
+      val payments =
+        (emp \ "payments" \ "inYear").validate[JsArray] match {
+          case JsSuccess(arr, _) =>
+            arr.value
+              .map { payment =>
+                payment.as[Payment](paymentHodReads)
+              }
+              .toList
+              .sorted
+          case _ => Nil
+        }
+
+      val eyus =
+        (emp \ "payments" \ "eyu").validate[JsArray] match {
+          case JsSuccess(arr, _) =>
+            arr.value
+              .map { payment =>
+                payment.as[EndOfTaxYearUpdate](endOfTaxYearUpdateHodReads)
+              }
+              .toList
+              .sorted
+          case _ => Nil
+        }
+
+      val taxYear = (json \ "individual" \ "relatedTaxYear").as[TaxYear](taxYearHodReads)
+
+      AnnualAccount(sequenceNumber, taxYear, Available, payments, eyus)
+    })
+  }
+
+  private implicit val stringMapFormat: Format[Map[String, BigDecimal]] =
+    JsonExtra.mapFormat[String, BigDecimal]("type", "amount")
 
 }
